Added `renv` capability
After a *lot* of fighting with the compiler, I added the capability to
record which packages are installed, and most importantly their version,
so that the same can be retrieved at a later date.
Reproducibility ftw!

This could also be a stepping stone for using Docker. I also tried just
using Docker right away, but the compilation takes just too much time.

The installation is made ergonomic through the `install_dep` script.