Moved logging to global option.
The logging option is shared by almost all commands. So, it was moved
out of the single commands. This also assures that we can call
`log_info` and the like wherever we please in the package without
having to care where we are writing to.

Several help messages were also updated. Importantly, `prepagil`
is now mentioned in the README and `GATTACA --help`.