Many bugs fixed:
1]
gsub(" ", "_", date()) returned this data format
"Fri_Oct_08_12:17:20_2021"
which is not a valid filename in Windows (because of the colon).
If GATTACA run in Docker, when looking at the final file from
Windows, : was substututed with a central dot, and that was fine.
However, in interactive mode (useful for debugging), that was an
issue. Now colon is avoided.

----------

2]
When using names(), the sussetting doesn't work.
names(user_colours[1:length(unique_simple_groups)]) <- unique_simple_groups
Its argument must be the entire vector, otherwise the reassigment will fail.

----------

3]
The number of detected groups is now logged.
This is useful to spot possible typos by user in the option file.

----------

4]
Colors in boxplot now match with group name (color coding restored)

----------

5]
Pre-DEA MA-Plot was flat because all differences were == 0,
because it was group1 == group2,
because of a wrong syntax in the pipe... no fixed:
expression_set |> dp_select(starts_with(combo[1])) |> rowMeans() -> group1

----------

6]
This was not a bug, but in hclust() 'ward.D2' is temporarily restored to
'ward.D' for backward comparison

----------

7]
In SD_vs_Mean Plot the 'Global' plot wasn't drawn... now it's fixed.

----------

8]
In filter section group's labels were switched when
rownames(..filtering_table_report) = unique_simple_groups
because 'unique_simple_groups' follows user-input order,
while '..filtering_table_report' does not

----------

9]
The same was true during '..limma_design' definition.
Now contrast's order is fixed by enforcing alphabetical order of colnames.
Only checked in unpaired-design mode!!!