Added MA plots to prepagil and prepaffy
The MA plots are made with a local implementation, which is prettier
than the rafalib or oligo or limma ones. Added relevant options to the
interfaces.

This commit also sets the stage for the calculation of the MA plot
"badness" score that can be used to sort the plots to show only the
top-n bad ones.

To solve the agilent probe number problem, the dataset is sampled with a
fair, fast algorithm, adapted from wikipedia. See the relevant
docstrings.