Added plot ranking capability
Plot ranking is done by fitting a smooth GAM, then binning the resulting
predictions and taking the average prediction values. This allows to
appreciate the distortion of the data.

The new `--plot-number` option allows to save to disk only the top `n`
plots. However, they still need to be computed beforehand (a possible
improvement).