Plot size fix and faster MA plot ordering
The previous way we computed the MA plot order required building the
plot to extract data from it. This caused building the plot twice, once
when making the order, and again when plotting. Now we calculate the
plot score while in the `get_better_mas` call, so the rebuilding is not
needed. This speeds the process up immensly.

The `--plot-size` option did not have an effect on the heigth of the
plots due to a typo. This has been corrected now.