Added new way to call the GATTACA entry script
The user-facing GATTACA script has to be able to change with the new
versions of the docker, to allow the calls to start the docker
appropriate for the various versions.

However, the GATTACA script has (had) two main flaws:
	1- The user had to manually download the proper version of the
	script from the GitHub release page to be sure to use the
	version of the script for their chosen GATTACA version, as the
	latest script (the one that one installs following the README)
	could be *too new* to properly run every container version.
	2- The script had no way to notify the user of a new version of
	the script itself, nor a way to auto-update.

This commit fixes both issues. Now, the GATTACA script merely queries
the docker for the contents of the `cli.sh` file (which is identical to
the old front-facing GATTACA file), and evaluates them. This allows the
user-facing GATTACA script to never change, while changes to the cli are
"version controlled" by the container itself, just like the rest of the
code. So, we get the flexibility and self-update capacity of the docker
container also for the cli script.

The only variable needed by the outside script is the `--version` of
the docker container. For semplicity and security's sake, this was
harcoded to be the first parameter that can be parsed by the script.
The call fails if this requirement is not met.

All old tests pass, with the change of moving the `--version` parameters
at the start of the call, as per above.

It migth not be the best solution, but I'm pretty proud of this one!