`prepAffy` does not crash if the platform is unrecognized.
PrepAffy used to stop if the platform was not recogized as it needs the
exon.probes flag to be set. Instead of stopping (which seriously limits
the types of data that can be processed) it asks the user what the flag
should be, giving a brief explaination.

The publish_version script pushed wrongly tagged images to docker hub
(namely, it did not strip the leading `v` to `v0.0.0` before tagging the
image).